// types/Results.ts
export type Results = {
  wavelength: number;
  spacingMM: number;
  circumferenceMM: number;
  diameter: number;
  axialLength: number;
  pitchAngle: number;
  turnLength: number;
  conductorLength: number;
  conductorDiameter: number;
  reflectorMin: number;
  HPBW: number;
  BWFN: number;
  gain: number;
  impedance: number;
};